refactor: migrate from FP16 to BF16 native support across all kernels

+++ src/engine/kernel/mlp/silu_mul.h
@@ -1,17 +1,17 @@
 #pragma once
 
-#include <cstdint>
-
-#include <cuda_fp16.h>
+#include <cuda_bf16.h>
 #include <cuda_runtime.h>
+
+#include <cstdint>
 
 #include "common/result.h"
 
 namespace ccinfer {
 namespace engine {
 
-Result<void> launch_silu_mul(const half* gate, const half* up, half* output, int64_t n,
-                              cudaStream_t stream);
+Result<void> launch_silu_mul(const __nv_bfloat16* gate, const __nv_bfloat16* up,
+                             __nv_bfloat16* output, int64_t n, cudaStream_t stream);
 
 }  // namespace engine
 }  // namespace ccinfer
